All pipeline stages complete I completed all the pipeline stages.  I haven't tested it, so I imagine there's probably a few bugs.  Additionally, I didn't implement jumping or stalling the pipeline yet.  So right now, the pipeline is suceptible to both RAW and WAR hazards and cannot halt.
I also moved a bunch of the memory interface into the control unit, as
it makes more sense to have one unified control unit to send read/write
signals.  These get passed on a 6 bit bus through the different stages
of the pipeline.  I also haven't implemented the cmp instruction, and
plan on having a 1 bit register to write the cmp result to in the
Instruction Decode stage.  This will should allow jumps to avoid
stalling the pipeline, as as soon as the jump enters the Instruction
Decode stage, the cmp bit will already be set, so it can immediately
change the program counter for the IF stage by muxing it with the jump
address.  The clock should be slow enough to allow this.